fix other controllers

+++ src/main/java/app/api/entity/SiteId.java
@@ -1,4 +1,4 @@
 package app.api.entity;
 
-public record SiteId(int siteId) {
+public record SiteId(Long siteId) {
 }
